Break out binary/lib resolution to new package for wider use
